requirements.txt
moviepy
fitparse
pandas
Pillow
numpy
smopy

